+++ demo/node/src/tests/runtime_api_mock.rs
@@ -201,6 +201,6 @@ pub fn mock_header() -> <Block as BlockT>::Header {
 		Default::default(),
 		Default::default(),
 		Default::default(),
-		Digest { logs: McHashInherentDigest::from_mc_block_hash(McBlockHash([1; 32])) },
+		Digest { logs: McHashInherentDigest::from_mc_block_hash(McBlockHash([0; 32])) },
 	)
 }